automate-oracle-forms-gemini/v2/requirements.txt
pyautogui
opencv-python-headless
pytesseract
numpy
faker
pytest
allure-pytest
Flask
